refactor: remove deprecated sales components #5
