+++ frontend/src/apps/sales/pages/index.tsx
@@ -1,0 +1,5 @@
+export { SalesDashboard } from './SalesDashboard';
+export { SalesList } from './SalesList';
+export { SaleForm } from './SaleForm';
+export { SaleDetail } from './SaleDetail';
+export { QuickSale } from './QuickSale';
